requirements.txt
alembic==1.14.0
amqp==5.3.1
aniso8601==9.0.1
bcrypt==4.0.1
billiard==4.2.1
blinker==1.7.0
cachelib==0.9.0
celery==5.4.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.0
Flask-Caching==2.3.0
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.0.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.4.2
lml==0.1.0
Mako==1.3.6
MarkupSafe==2.1.3
passlib==1.7.4
prompt_toolkit==3.0.48
pycparser==2.21
pyexcel==0.7.1
pyexcel-io==0.6.7
pyexcel-webio==0.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
requests==2.32.3
six==1.16.0
soundfile==0.12.1
SQLAlchemy==2.0.23
texttable==1.7.0
typing_extensions==4.8.0
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.2.1
